requirements.txt
django~=5.1.2
django-allauth
pillow
requests
requests-oauthlib
python3-openid
pyjwt
python3-saml
channels~=4.1.0
django-debug-toolbar~=4.4.6
asgiref~=3.8.1
django-debug-toolbar
